health-robot-backend/requirements.txt
fastapi
uvicorn
paho-mqtt
pydantic
aiofiles
sqlite-utils
requests
